moved collector creator for refactoring
